dev_environment/orchestration/requirements.txt
# Composer DAG Requirements
# Additional packages needed for the Composer environment

# Core data processing
pandas-gbq>=0.26.1
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
pyarrow>=12.0.0

# Database connectivity
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
google-auth>=2.40.3
